+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -9,6 +9,7 @@ import { EventEmitter } from 'events'
 import getPort from 'get-port'
 import PeerId from 'peer-id'
 import { removeFilesFromDir } from '../common/utils'
+
 import {
   AskForMessagesPayload,
   ChannelMessagesIdsResponse,
@@ -59,6 +60,7 @@ import { StorageEvents } from '../storage/storage.types'
 import { LazyModuleLoader } from '@nestjs/core'
 import Logger from '../common/logger'
 import { emitError } from '../socket/socket.errors'
+import { isPSKcodeValid } from '@quiet/common'
 
 @Injectable()
 export class ConnectionsManagerService extends EventEmitter implements OnModuleInit {
@@ -155,7 +157,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     this.logger('launchCommunityFromStorage')
 
     const community: InitCommunityPayload = await this.localDbService.get(LocalDBKeys.COMMUNITY)
-    console.log('launchCommunityFromStorage - community', community)
+    this.logger('launchCommunityFromStorage - community:', community?.id)
     if (community) {
       const sortedPeers = await this.localDbService.getSortedPeers(community.peers)
       if (sortedPeers.length > 0) {
@@ -164,8 +166,6 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       await this.localDbService.put(LocalDBKeys.COMMUNITY, community)
       if ([ServiceState.LAUNCHING, ServiceState.LAUNCHED].includes(this.communityState)) return
       this.communityState = ServiceState.LAUNCHING
-    }
-    if (community) {
       await this.launchCommunity(community)
     }
   }
@@ -177,6 +177,8 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     if (this.storageService) {
       this.logger('Stopping orbitdb')
       await this.storageService?.stopOrbitDb()
+      this.logger('reset CsrReplicated map and id')
+      this.storageService.resetCsrReplicatedMapAndId()
     }
     // if (this.storageService.ipfs) {
     //   this.storageService.ipfs = null
@@ -274,18 +276,43 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       },
       network,
     }
+    const psk = community.psk
+    if (psk) {
+      this.logger('Creating network: received Libp2p PSK')
+      if (!isPSKcodeValid(psk)) {
+        this.logger.error('Creating network: received Libp2p PSK is not valid')
+        emitError(this.serverIoProvider.io, {
+          type: SocketActionTypes.NETWORK,
+          message: ErrorMessages.NETWORK_SETUP_FAILED,
+          community: community.id,
+        })
+        return
+      }
+      await this.localDbService.put(LocalDBKeys.PSK, psk)
+    }
+
     this.serverIoProvider.io.emit(SocketActionTypes.NETWORK, payload)
   }
 
+  private async generatePSK() {
+    const pskBase64 = Libp2pService.generateLibp2pPSK().psk
+    await this.localDbService.put(LocalDBKeys.PSK, pskBase64)
+    this.logger('Generated Libp2p PSK')
+    this.serverIoProvider.io.emit(SocketActionTypes.LIBP2P_PSK_SAVED, { psk: pskBase64 })
+  }
+
   public async createCommunity(payload: InitCommunityPayload) {
-    console.log('ConnectionsManager.createCommunity peers:', payload.peers)
+    this.logger('Creating community: peers:', payload.peers)
+
+    await this.generatePSK()
+
     await this.launchCommunity(payload)
     this.logger(`Created and launched community ${payload.id}`)
     this.serverIoProvider.io.emit(SocketActionTypes.NEW_COMMUNITY, { id: payload.id })
   }
 
   public async launchCommunity(payload: InitCommunityPayload) {
-    console.log('ConnectionsManager.launchCommunity peers:', payload.peers)
+    this.logger('Launching community: peers:', payload.peers)
     this.communityState = ServiceState.LAUNCHING
     const communityData: InitCommunityPayload = await this.localDbService.get(LocalDBKeys.COMMUNITY)
     if (!communityData) {
@@ -329,7 +356,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       targetPort: this.ports.libp2pHiddenService,
       privKey: payload.hiddenService.privateKey,
     })
-    this.logger(`Launching community ${payload.id}, peer: ${payload.peerId.id}`)
+    this.logger(`Launching community ${payload.id}: peer: ${payload.peerId.id}`)
 
     const { Libp2pModule } = await import('../libp2p/libp2p.module')
     const moduleRef = await this.lazyModuleLoader.load(() => Libp2pModule)
@@ -341,11 +368,17 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     const _peerId = await peerIdFromKeys(restoredRsa.marshalPubKey(), restoredRsa.marshalPrivKey())
 
     let peers = payload.peers
-    console.log(`Launching community ${payload.id}, payload peers: ${peers}`)
+    this.logger(`Launching community ${payload.id}: payload peers: ${peers}`)
     if (!peers || peers.length === 0) {
       peers = [this.libp2pService.createLibp2pAddress(onionAddress, _peerId.toString())]
     }
+    const pskValue: string = await this.localDbService.get(LocalDBKeys.PSK)
+    if (!pskValue) {
+      throw new Error('No psk in local db')
+    }
+    this.logger(`Launching community ${payload.id}: retrieved Libp2p PSK`)
 
+    const libp2pPSK = Libp2pService.generateLibp2pPSK(pskValue).fullKey
     const params: Libp2pNodeParams = {
       peerId: _peerId,
       listenAddresses: [this.libp2pService.createLibp2pListenAddress(onionAddress)],
@@ -353,16 +386,15 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       localAddress: this.libp2pService.createLibp2pAddress(onionAddress, _peerId.toString()),
       targetPort: this.ports.libp2pHiddenService,
       peers,
+      psk: libp2pPSK,
     }
 
     await this.libp2pService.createInstance(params)
-    // KACPER
     // Libp2p event listeners
     this.libp2pService.on(Libp2pEvents.PEER_CONNECTED, (payload: { peers: string[] }) => {
       this.serverIoProvider.io.emit(SocketActionTypes.PEER_CONNECTED, payload)
     })
     this.libp2pService.on(Libp2pEvents.PEER_DISCONNECTED, async (payload: NetworkDataPayload) => {
-      console.log(' this.libp2pService.on(Libp2pEvents.PEER_DISCONNECTED')
       const peerPrevStats = await this.localDbService.find(LocalDBKeys.PEERS, payload.peer)
       const prev = peerPrevStats?.connectionTime || 0
 
@@ -379,7 +411,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       this.serverIoProvider.io.emit(SocketActionTypes.PEER_DISCONNECTED, payload)
     })
     await this.storageService.init(_peerId)
-    console.log('storage initialized')
+    this.logger('storage initialized')
   }
   private attachTorEventsListeners() {
     this.logger('attachTorEventsListeners')
@@ -397,6 +429,12 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     })
     this.registrationService.on(RegistrationEvents.NEW_USER, async payload => {
       await this.storageService?.saveCertificate(payload)
+    })
+
+    this.registrationService.on(RegistrationEvents.FINISHED_ISSUING_CERTIFICATES_FOR_ID, payload => {
+      if (payload.id) {
+        this.storageService.resolveCsrReplicatedPromise(payload.id)
+      }
     })
   }
   private attachsocketServiceListeners() {
@@ -420,6 +458,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       }
     })
     this.socketService.on(SocketActionTypes.CREATE_NETWORK, async (args: Community) => {
+      this.logger(`socketService - ${SocketActionTypes.CREATE_NETWORK}`)
       await this.createNetwork(args)
     })
     this.socketService.on(SocketActionTypes.CREATE_COMMUNITY, async (args: InitCommunityPayload) => {
@@ -443,7 +482,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       await this.storageService?.updateCommunityMetadata(payload)
     })
     this.socketService.on(SocketActionTypes.SAVE_USER_CSR, async (payload: SaveCSRPayload) => {
-      console.log(`On ${SocketActionTypes.SAVE_USER_CSR}`)
+      this.logger(`socketService - ${SocketActionTypes.SAVE_USER_CSR}`)
       await this.storageService?.saveCSR(payload)
       this.serverIoProvider.io.emit(SocketActionTypes.SAVED_USER_CSR, payload)
     })
@@ -492,7 +531,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     this.socketService.on(
       SocketActionTypes.DELETE_FILES_FROM_CHANNEL,
       async (payload: DeleteFilesFromChannelSocketPayload) => {
-        this.logger('DELETE_FILES_FROM_CHANNEL : payload', payload)
+        this.logger(`socketService - ${SocketActionTypes.DELETE_FILES_FROM_CHANNEL}`, payload)
         await this.storageService?.deleteFilesFromChannel(payload)
         // await this.deleteFilesFromTemporaryDir() //crashes on mobile, will be fixes in next versions
       }
@@ -525,7 +564,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       this.serverIoProvider.io.emit(SocketActionTypes.CHANNEL_SUBSCRIBED, payload)
     })
     this.storageService.on(StorageEvents.CREATED_CHANNEL, (payload: CreatedChannelResponse) => {
-      console.log('created channel in services')
+      this.logger(`Storage - ${StorageEvents.CREATED_CHANNEL}: ${payload.channel.name}`)
       this.serverIoProvider.io.emit(SocketActionTypes.CREATED_CHANNEL, payload)
     })
     this.storageService.on(StorageEvents.REMOVE_DOWNLOAD_STATUS, (payload: RemoveDownloadStatus) => {
@@ -556,19 +595,19 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       this.serverIoProvider.io.emit(SocketActionTypes.CHECK_FOR_MISSING_FILES, payload)
     })
     this.storageService.on(StorageEvents.CHANNEL_DELETION_RESPONSE, (payload: { channelId: string }) => {
-      console.log('emitting deleted channel event back to state manager')
+      this.logger(`Storage - ${StorageEvents.CHANNEL_DELETION_RESPONSE}`)
       this.serverIoProvider.io.emit(SocketActionTypes.CHANNEL_DELETION_RESPONSE, payload)
     })
     this.storageService.on(
       StorageEvents.REPLICATED_CSR,
-      async (payload: { csrs: string[]; certificates: string[] }) => {
-        console.log(`On ${StorageEvents.REPLICATED_CSR}`)
+      async (payload: { csrs: string[]; certificates: string[]; id: string }) => {
+        console.log(`Storage - ${StorageEvents.REPLICATED_CSR}`)
         this.serverIoProvider.io.emit(SocketActionTypes.RESPONSE_GET_CSRS, { csrs: payload.csrs })
         this.registrationService.emit(RegistrationEvents.REGISTER_USER_CERTIFICATE, payload)
       }
     )
     this.storageService.on(StorageEvents.REPLICATED_COMMUNITY_METADATA, (payload: CommunityMetadata) => {
-      console.log(`On ${StorageEvents.REPLICATED_COMMUNITY_METADATA}: ${payload}`)
+      this.logger(`Storage - ${StorageEvents.REPLICATED_COMMUNITY_METADATA}: ${payload}`)
       const communityMetadataPayload: CommunityMetadataPayload = {
         rootCa: payload.rootCa,
         ownerCertificate: payload.ownerCertificate,