+++ packages/desktop/src/renderer/components/widgets/possibleImpersonationAttackModal/PossibleImpersonationAttackModal.test.tsx
@@ -8,12 +8,7 @@ describe('PossibleImpersonationAttackModal', () => {
   it('renderComponent', () => {
     const result = renderComponent(
       <ThemeProvider theme={theme}>
-        <PossibleImpersonationAttackModalComponent
-          handleClose={() => {}}
-          open={true}
-          communityName={'devteam'}
-          leaveCommunity={() => {}}
-        />
+        <PossibleImpersonationAttackModalComponent handleClose={() => {}} open={true} communityName={'devteam'} />
       </ThemeProvider>
     )
     expect(result.baseElement).toMatchInlineSnapshot(`
@@ -121,17 +116,6 @@ describe('PossibleImpersonationAttackModal', () => {
                       This should never happen and we recommend leaving this community immediately!
                     </strong>
                   </p>
-                  <button
-                    class="MuiButtonBase-root MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium MuiButton-root MuiButton-text MuiButton-textPrimary MuiButton-sizeMedium MuiButton-textSizeMedium PossibleImpersonationAttackModalComponent-button css-1skytee-MuiButtonBase-root-MuiButton-root"
-                    data-testid="unregistered-button"
-                    tabindex="0"
-                    type="button"
-                  >
-                    Leave community
-                    <span
-                      class="MuiTouchRipple-root css-8je8zh-MuiTouchRipple-root"
-                    />
-                  </button>
                 </div>
               </div>
             </div>