+++ packages/desktop/src/renderer/index.tsx
@@ -6,6 +6,7 @@ import Root, { persistor } from './Root'
 import store from './store'
 import updateHandlers from './store/handlers/update'
 import { communities } from '@quiet/state-manager'
+import { InvitationData } from '@quiet/types'
 
 if (window && process.env.DEBUG) {
   window.localStorage.setItem('debug', process.env.DEBUG)
@@ -20,9 +21,10 @@ ipcRenderer.on('force-save-state', async _event => {
   ipcRenderer.send('state-saved')
 })
 
-ipcRenderer.on('invitation', (_event, invitation) => {
-  console.log('invitation', invitation, 'dispatching action')
-  store.dispatch(communities.actions.handleInvitationCodes(invitation.codes))
+ipcRenderer.on('invitation', (_event, invitation: { data: InvitationData }) => {
+  if (!invitation.data) return
+  console.log('invitation', invitation.data.pairs, 'dispatching action')
+  store.dispatch(communities.actions.customProtocol(invitation.data))
 })
 
 const container = document.getElementById('root')